Fix a specification implementation issue relating to creating the correct amount of jobs on order request.
The CESSDA Café specifications require an individual job for each coffee. Previously the cashier combined multiple requests for the same coffee into the same job, and used the orderSize variable to store the amount of coffees of each type that had been ordered in the job.

The correct implementation was to create an individual job per coffee and set the orderSize in the job to the total size of the order. This has now been implemented.